Static-analysis pass: cppcheck clean; fix real clang-tidy findings
cppcheck (host code) reported zero findings. clang-tidy (bugprone/performance/misc) findings
triaged — fixed the genuinely-real ones, bit-identical (verified single_gaussian + cloud):
- image.h total_size(): cast first operand to size_t before the width*height*spp multiply —
  was a uint32 multiplication widened afterward (latent overflow for large/high-spp images;
  pixel_count() already did this).
- context.h: MEGABYTES uses a size_t literal (was an int*int widened to size_t).
- result.h: parenthesize the TRY_ASSIGN macro's `var` argument (macro hygiene).
- test_runner.cpp: run_test_scene() output_path → string_view (CLAUDE.md read-only param).

The remaining clang-tidy notes (easily-swappable-parameters, non-private members on POD/struct
types, const-correctness on locals, enum-size, exception-escape from CLI setup) are low-value
style or intentional design and were left as-is.

+++ include/thesis/host/params/image.h
@@ -123,7 +123,8 @@ class Image {
         return static_cast<size_t>(device_image_.width_) * device_image_.height_;
     }
     [[nodiscard]] size_t total_size() const noexcept {
-        return device_image_.width_ * device_image_.height_ * device_image_.num_samples_per_pixel_;
+        return static_cast<size_t>(device_image_.width_) * device_image_.height_ *
+               device_image_.num_samples_per_pixel_;
     }
 
     [[nodiscard]] float aspect_ratio() const noexcept {
